Installed alembic as the database version control system.
