+++ backend/mms_app_backend/mms_app_backend/configs/alembic_config/versions/5f28072965c1_alembic_initial_revision.py
@@ -1,0 +1,28 @@
+"""Alembic initial revision
+
+Revision ID: 5f28072965c1
+Revises: 
+Create Date: 2023-04-20 22:28:50.430830
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '5f28072965c1'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
